One small change, and a (I believe) finished project
I changed on the spelling of initial from "inital" to "initial" on: contract.html, index.html, introduction.html, review.html, and tables.html.

I also fixed the css problems that were apparent to me. This was almost all if not all inside of the the media queries and had to do with list of different types.